Initialize screenshot testing and change to detekt compose
