+++ app/src/androidTest/kotlin/com/abnamro/apps/referenceandroid/screenshotTest/ScreenshotTest.kt
@@ -1,0 +1,18 @@
+package com.abnamro.apps.referenceandroid.screenshotTest
+
+import android.R
+import android.view.View
+
+
+class ScreenshotTest {
+//    @Test
+//    fun doScreenshot() {
+//        val view: View? = mLayoutInflater.inflate(R.layout., null, false)
+//        ViewHelpers.setupView(view)
+//            .setExactWidthDp(300)
+//            .layout()
+//
+//        Screenshot.snap(view)
+//            .record()
+//    }
+}
