--- app/src/main/java/com/abnamro/apps/referenceandroid/MainActivityFragment.kt
@@ -1,20 +1,0 @@
-package com.abnamro.apps.referenceandroid
-
-import androidx.fragment.app.Fragment
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-/**
- * A placeholder fragment containing a simple view.
- */
-class MainActivityFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_main, container, false)
-    }
-}
