--- app/src/main/java/com/abnamro/apps/referenceandroid/screens/CommentsListScreen.kt
@@ -1,196 +1,0 @@
-package com.abnamro.apps.referenceandroid.screens
-
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Refresh
-import androidx.compose.material.icons.filled.Warning
-import androidx.compose.material3.Card
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SnackbarHost
-import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.compose.LocalLifecycleOwner
-import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.repeatOnLifecycle
-import com.abnamro.apps.referenceandroid.model.Comment
-import com.abnamro.apps.referenceandroid.viewmodel.CommentsViewModel
-import org.koin.androidx.compose.koinViewModel
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CommentsListScreen(
-    viewModel: CommentsViewModel
-) {
-    val snackBarHostState = remember { SnackbarHostState() }
-    val commentsUIState by viewModel.commentsUIState.collectAsStateWithLifecycle()
-
-    LaunchedEffect(Unit) {
-        viewModel.uiEvent.collect { event ->
-            when (event) {
-                is CommentsViewModel.UIEvent.ShowSnackBar -> {
-                    snackBarHostState.showSnackbar(event.message)
-                }
-            }
-        }
-    }
-
-    Scaffold(topBar = {
-        TopAppBar(
-            title = { Text(text = "Comments") }, colors = TopAppBarDefaults.topAppBarColors(
-                containerColor = MaterialTheme.colorScheme.primary,
-                titleContentColor = MaterialTheme.colorScheme.onPrimary,
-            )
-        )
-    }, floatingActionButton = {
-        FloatingActionButton(
-            onClick = {
-                if (!commentsUIState.isLoading) {
-                    viewModel.getComments()
-                }
-            },
-        ) {
-            Icon(imageVector = Icons.Default.Refresh, contentDescription = "Refresh comment")
-        }
-    }, snackbarHost = {
-        SnackbarHost(hostState = snackBarHostState)
-    }, content = { paddingValues ->
-        Column(
-            Modifier
-                .fillMaxSize()
-                .padding(paddingValues),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            AnimatedVisibility(visible = commentsUIState.comments.isNotEmpty()) {
-                LazyColumn {
-                    items(items = commentsUIState.comments) { comment ->
-                        CommentItem(comment = comment)
-                    }
-                }
-            }
-
-            AnimatedVisibility(
-                visible = (commentsUIState.comments.isEmpty() && !commentsUIState.isLoading)
-            ) {
-                Text("Please press the button to load comments", fontSize = 16.sp)
-            }
-
-            AnimatedVisibility(visible = commentsUIState.isLoading) {
-                CircularProgressIndicator()
-            }
-
-            AnimatedVisibility(visible = commentsUIState.error != null) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.Warning,
-                        contentDescription = "Error",
-                        tint = Color.Red,
-                        modifier = Modifier
-                            .size(60.dp)
-                            .padding(bottom = 10.dp)
-                    )
-                    Text(
-                        text = "There was an error! Try again...",
-                        color = Color.Red,
-                        fontSize = 16.sp
-                    )
-                }
-            }
-        }
-    })
-}
-
-
-@Composable
-fun CommentItem(comment: Comment) {
-    Card(
-        modifier = Modifier
-            .padding(top = 10.dp, start = 10.dp, end = 10.dp)
-            .fillMaxWidth(),
-        elevation = CardDefaults.cardElevation(4.dp)
-    ) {
-        Column(
-            modifier = Modifier.padding(16.dp)
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier.padding(bottom = 8.dp)
-            ) {
-                Box(
-                    modifier = Modifier
-                        .size(40.dp)
-                        .clip(CircleShape)
-                        .background(MaterialTheme.colorScheme.primary),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = comment.email.first().uppercaseChar().toString(),
-                        color = MaterialTheme.colorScheme.onPrimary,
-                        fontWeight = FontWeight.Bold,
-                        fontSize = 24.sp
-                    )
-                }
-                Spacer(modifier = Modifier.width(8.dp))
-                Column {
-                    Text(
-                        text = comment.name,
-                        style = MaterialTheme.typography.titleMedium,
-                        fontWeight = FontWeight.Bold
-                    )
-                    Text(
-                        text = comment.email,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                }
-            }
-            Text(
-                text = comment.body,
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier.padding(top = 8.dp)
-            )
-        }
-    }
-}
-
