+++ app/src/main/kotlin/com/abnamro/apps/referenceandroid/network/NetworkResult.kt
