+++ app/src/screenshotTest/kotlin/com/abnamro/apps/referenceandroid/ExamplePreviewScreenshotTest.kt
@@ -1,0 +1,4 @@
+package com.abnamro.apps.referenceandroid
+
+class ExamplePreviewScreenshotTest {
+}
