+++ build.gradle
@@ -6,10 +6,12 @@ buildscript {
         coroutines_version = '1.10.2'
         koin_version = '4.1.0'
     }
+
     repositories {
         google()
         mavenCentral()
     }
+
     dependencies {
         classpath 'com.android.tools.build:gradle:8.11.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
@@ -18,9 +20,8 @@ buildscript {
 }
 
 plugins {
-    id "io.gitlab.arturbosch.detekt" version "1.23.8"
     id "org.sonarqube" version "6.2.0.5505"
-    id 'org.jetbrains.kotlin.plugin.compose' version '2.2.0' apply false
+    id "org.jetbrains.kotlin.plugin.compose" version "2.2.0" apply false
 }
 
 allprojects {
@@ -30,22 +31,12 @@ allprojects {
     }
 }
 
-subprojects {
-    apply plugin: "io.gitlab.arturbosch.detekt"
-    detekt {
-        buildUponDefaultConfig = true
-        allRules = false
-        config = files("${project.rootDir}/config/detekt/detekt.yml")
-        parallel = true
-    }
+tasks.register("kotlinLintCheck") {
+    dependsOn("detekt")
+    group = "verification"
 }
 
-tasks.register('kotlinLintCheck') {
-    dependsOn['detekt']
-    group = 'verification'
-}
-
-tasks.register('clean') {
+tasks.register("clean") {
     delete getLayout().getBuildDirectory()
 }
 
@@ -55,15 +46,14 @@ sonar {
         property "sonar.login", "1c33e69cccc01ffafc22414db270ae2f172bfd70"
         property "sonar.projectName", "BUUT Assessment Application"
         property "sonar.projectKey", "buut-assessment-app"
+        property "sonar.projectVersion", "1.0.0"
+        property "sonar.description", "This is the project for BUUT assessment"
         property "sonar.sourceEncoding", "UTF-8"
         property "sonar.sources", "./src/main"
-        property "sonar.description", "This is the project for BUUT assessment"
         property "sonar.report.export.path", "sonar-report.json"
-        property "sonar.projectVersion","1.0.0"
-        property "detekt.sonar.kotlin.config.path","config/detekt/detekt.yml"
+        property "detekt.sonar.kotlin.config.path", "config/detekt/detekt.yml"
         property "sonar.android.lint.report", "./build/reports/lint-results-debug.xml"
-        property 'sonar.core.codeCoveragePlugin', 'jacoco'
-        property 'sonar.coverage.jacoco.xmlReportPaths', "app/build/reports/coverage/androidTest/debug/connected/report.xml"
+        property "sonar.core.codeCoveragePlugin", "jacoco"
+        property "sonar.coverage.jacoco.xmlReportPaths", "app/build/reports/coverage/androidTest/debug/connected/report.xml"
     }
 }
-
