+++ gradle.properties
@@ -16,3 +16,4 @@ kotlin.jvm.target.validation.mode = IGNORE
 kotlin.code.style=official
 android.useAndroidX=true
 android.enableJetifier=true
+android.experimental.enableScreenshotTest=true
